requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic
streamlit
requests
plotly
python-multipart